requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-experimental
langchain-neo4j
neo4j
python-dotenv
tiktoken
supabase
fastapi
uvicorn[standard]
twilio
httpx